+++ src/main/java/com/twilio/jwt/taskrouter/FilterRequirement.java
@@ -1,5 +1,8 @@
 package com.twilio.jwt.taskrouter;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+@JsonFormat(shape = JsonFormat.Shape.OBJECT)
 public enum FilterRequirement {
     REQUIRED(true),
     OPTIONAL(false);